Image_segmentation/requirements.txt
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
torch
###
albumentations
tqdm
torchvision